[app] androidx.navigation, MainActivity as HomeFragment
