+++ app/build.gradle
@@ -108,11 +108,14 @@ dependencies {
     final def glide_version = '4.12.0'
     final def markwon_version = '4.6.2'
     final def okhttp_version = '4.9.1'
+    final def nav_version = "2.3.3"
     annotationProcessor "com.github.bumptech.glide:compiler:$glide_version"
     implementation 'androidx.activity:activity:1.2.0'
     implementation 'androidx.browser:browser:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'androidx.fragment:fragment:1.3.0'
+    implementation "androidx.navigation:navigation-fragment:$nav_version"
+    implementation "androidx.navigation:navigation-ui:$nav_version"
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     implementation 'com.caverock:androidsvg-aar:1.4'
     implementation "com.github.bumptech.glide:glide:$glide_version"
