+++ app/src/main/java/io/github/lsposed/manager/util/chrome/CustomTabsURLSpan.java
@@ -20,17 +20,17 @@
 
 package io.github.lsposed.manager.util.chrome;
 
+import android.app.Activity;
 import android.text.style.URLSpan;
 import android.view.View;
 
-import io.github.lsposed.manager.ui.activity.base.BaseActivity;
 import io.github.lsposed.manager.util.NavUtil;
 
 public class CustomTabsURLSpan extends URLSpan {
 
-    private final BaseActivity activity;
+    private final Activity activity;
 
-    public CustomTabsURLSpan(BaseActivity activity, String url) {
+    public CustomTabsURLSpan(Activity activity, String url) {
         super(url);
         this.activity = activity;
     }
